Refactor clickable calendar views into absences module
