+++ ap_src/ap_app/urls.py
@@ -41,8 +41,8 @@ urlpatterns = [
     path("absence_recurring/delete/<int:absence_id>/<int:user_id>/<int:team_id>/", absences.absence_recurring_delete, name="recurring_absence_delete"),
     path("absence/edit/<int:pk>", absences.EditAbsence.as_view(), name="absence_edit"),
     path("absence/edit_recurring/<int:pk>", absences.edit_recurring_absences, name="recurring_absence_edit"),
-    path("absence/click_add", views.click_add, name="absence_click_add"),
-    path("absence/click_remove", views.click_remove, name="absence_click_remove"),
+    path("absence/click_add", absences.click_add, name="absence_click_add"),
+    path("absence/click_remove", absences.click_remove, name="absence_click_remove"),
     path("profile/settings", views.profile_settings, name="profile_settings"),
     path("profile/settings/add-user", views.add_user, name="add-user"),
     path("profile/settings/set-region", views.set_region, name="set-region"),
